requirements.txt
boto3
botocore
ipykernel
pandas
pymupdf
google-generativeai